Separate the code for GET /ingests and POST /ingests
This matches the refactoring already done in the bags API.
